refactor: add UserSerializerTest for validating UserSerializer functionality
